% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lux_spectrum.R
\name{convert_unit}
\alias{convert_unit}
\title{Convert a lux_spectrum to a different unit}
\usage{
convert_unit(x, to, ...)
}
\arguments{
\item{x}{A \code{lux_spectrum} object.}

\item{to}{Target unit string from the controlled vocabulary.}

\item{...}{Ignored.}
}
\value{
A new \code{lux_spectrum} with updated \code{unit} and \code{E}.
}
\description{
Converts the spectral values in a \code{lux_spectrum} between energy and
photon-flux units, or between mW and W for radiance spectra. Supported
conversions: \code{"W/m2/nm"} \eqn{\leftrightarrow}
\code{"umol/m2/s/nm"} / \code{"mmol/m2/s/nm"} / \code{"mol/m2/s/nm"},
and the corresponding \code{"/sr"} photonic units. Energy radiance units
\code{"W/m2/sr/nm"} and \code{"mW/m2/sr/nm"} are also interchangeable.
}
\examples{
lam   <- seq(400, 700, by = 10)
x_W   <- lux_spectrum(rep(1, length(lam)), lam, "irradiance", "W/m2/nm")
x_mol <- convert_unit(x_W, "umol/m2/s/nm")
x_mol$unit
}
