% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum_readers.R
\name{from_naples}
\alias{from_naples}
\title{Load a Naples depth spectrum as a lux_spectrum}
\usage{
from_naples(depth = c("0m", "5m", "10m"))
}
\arguments{
\item{depth}{One of \code{"0m"}, \code{"5m"}, \code{"10m"}.}
}
\value{
A \code{lux_spectrum} with unit \code{"umol/m2/s/nm"}.
}
\description{
Load a Naples depth spectrum as a lux_spectrum
}
\examples{
x <- from_naples("0m")
print(x)
}
\seealso{
\code{\link{Naples}}, \code{\link{as_lux_spectrum}}
}
