% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mol2photon_FUN.R
\name{mol2photon_FUN}
\alias{mol2photon_FUN}
\title{Converts photon number in moles or similar to n.}
\usage{
mol2photon_FUN(value, molar_unit = NULL)
}
\arguments{
\item{value}{The value to be converted to photon number.}

\item{molar_unit}{The metric used, e.g. moles or micromoles.}
}
\value{
The value in photons.
}
\description{
Converts photon number in moles or similar to n.
}
\keyword{internal}
