% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polarization.R
\name{scattering_angle}
\alias{scattering_angle}
\title{Angle between a line of sight and the solar beam}
\usage{
scattering_angle(view_zenith, view_azimuth, sun_zenith, sun_azimuth)
}
\arguments{
\item{view_zenith, view_azimuth}{Viewing direction in degrees.}

\item{sun_zenith, sun_azimuth}{Solar-beam direction in degrees (use
\code{\link{refracted_solar_angle}} for the in-water zenith).}
}
\value{
Angle in degrees, \code{[0, 180]}. Vectorised over its arguments.
}
\description{
The angle \eqn{\Theta} between two directions given as (zenith, azimuth):
\deqn{\cos\Theta = \cos\theta_1\cos\theta_2 +
       \sin\theta_1\sin\theta_2\cos(\phi_1 - \phi_2).}
For the single-scattering polarization model this is the relevant scattering
angle (the degree of polarization peaks near \eqn{\Theta = 90^\circ}).
}
\examples{
scattering_angle(view_zenith = 90, view_azimuth = 0,
                 sun_zenith = 0,  sun_azimuth = 0)   # 90
}
\seealso{
\code{\link{underwater_polarization}}
}
