% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optical_properties.R
\name{single_scattering_albedo}
\alias{single_scattering_albedo}
\title{Single-scattering albedo.}
\usage{
single_scattering_albedo(absorption, scattering)
}
\arguments{
\item{absorption}{Absorption coefficient \eqn{a} (1/m). Vectorised.}

\item{scattering}{Scattering coefficient \eqn{b} (1/m). Vectorised.}
}
\value{
Single-scattering albedo \eqn{\omega_0} (dimensionless, 0-1).
}
\description{
The fraction of beam attenuation due to scattering rather than absorption,
\eqn{\omega_0 = b / (a + b) = b / c} (Johnsen 2012, ch. 5). Values near 1
indicate a strongly scattering, weakly absorbing medium (e.g. fog or milk);
values near 0 a strongly absorbing one.
}
\examples{
single_scattering_albedo(absorption = 0.1, scattering = 0.05)   # 0.333
}
\references{
Johnsen S (2012) The Optics of Life. Princeton University Press, ch. 5.
}
\seealso{
\code{\link{beam_attenuation}}
}
