% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optical_properties.R
\name{transmittance}
\alias{transmittance}
\title{Beer-Lambert transmittance over a path.}
\usage{
transmittance(coefficient, path_length)
}
\arguments{
\item{coefficient}{Attenuation (or absorption) coefficient (1/length).
Vectorised.}

\item{path_length}{Path length in the same length unit. Vectorised.}
}
\value{
Transmittance (0-1).
}
\description{
Fraction of collimated light transmitted through a path of length
\code{path_length} in a medium with attenuation \code{coefficient}:
\eqn{T = e^{-c\,\ell}} (Johnsen 2012, ch. 4). The fraction \emph{absorbed}
(or attenuated) is \eqn{1 - T}; note that absorbances add over stacked paths
while transmittances multiply.
}
\examples{
transmittance(coefficient = 0.15, path_length = 10)
1 - transmittance(0.064, 10)   # hatchetfish rod: ~47\% absorbed
}
\references{
Johnsen S (2012) The Optics of Life. Princeton University Press, ch. 4.
}
\seealso{
\code{\link{absorbance}}, \code{\link{beam_attenuation}}
}
