% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luxR-package.R
\docType{package}
\name{luxR-package}
\alias{luxR}
\alias{luxR-package}
\title{luxR: Light Exposure and Photic Regularity Analysis for Chronobiology Research}
\description{
Tools for processing ambient light exposure data collected from
wrist-worn actigraphy devices (Axivity AX3/AX6 and ActTrust) in the
context of circadian and chronobiology research: photic exposure metrics
(mean/median illuminance, time above threshold, cumulative photic
history, melanopic EDI approximation) and light timing/regularity
metrics (light onset/offset detection, mean light timing, Light
Regularity Index).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://lux.circadia-lab.uk}
  \item \url{https://github.com/circadia-bio/luxR}
  \item Report bugs at \url{https://github.com/circadia-bio/luxR/issues}
}

}
\author{
\strong{Maintainer}: Lucas França \email{lucas.franca@northumbria.ac.uk} (\href{https://orcid.org/0000-0003-0853-1319}{ORCID})

Authors:
\itemize{
  \item Lucas França \email{lucas.franca@northumbria.ac.uk} (\href{https://orcid.org/0000-0003-0853-1319}{ORCID})
  \item Mario Leocadio-Miguel (\href{https://orcid.org/0000-0002-7248-3529}{ORCID})
}

}
\keyword{internal}
