% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n2W_spec_irradiance_FUN.R
\name{n2W_spec_irradiance_FUN}
\alias{n2W_spec_irradiance_FUN}
\title{Provide W values for measures of spectral irradiance. The constant is Planck's constant times the speed of light in nm/s. It is an internal function applied to a single wavelength bin.}
\usage{
n2W_spec_irradiance_FUN(photons, lambda)
}
\arguments{
\item{photons}{The spectral irradiance (in \eqn{W\,m^{-2}\,nm^{-1}} or
\eqn{photons\,m^{-2}\,s^{-1}\,nm^{-1}}; see photon) at each wavelength bin.}

\item{lambda}{The wavelength in nanometres (nm), taken from the midpoint value of irradiance bins.}
}
\value{
A scalar representing the spectral irradiance values transformed
  from \eqn{photons\,m^{-2}\,s^{-1}\,nm^{-1}} into \eqn{W\,m^{-2}\,nm^{-1}}.
}
\description{
Provide W values for measures of spectral irradiance. The constant is Planck's constant times the speed of light in nm/s. It is an internal function applied to a single wavelength bin.
}
\keyword{internal}
