% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{W2photon}
\alias{W2photon}
\title{Convert watts to photon counts at a given wavelength.}
\usage{
W2photon(W, lambda)
}
\arguments{
\item{W}{Spectral irradiance in \eqn{W\,m^{-2}\,nm^{-1}}{W m^-2 nm^-1}.}

\item{lambda}{Wavelength(s) in nm.}
}
\value{
Numeric vector of photon counts in \eqn{photons\,m^{-2}\,s^{-1}\,nm^{-1}}{photons m^-2 s^-1 nm^-1}.
}
\description{
Inverse of \code{\link{photon2W}}.
}
\examples{
W2photon(1, 555)
}
