% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{W2photon}
\alias{W2photon}
\title{Convert watts to photon counts at a given wavelength.}
\usage{
W2photon(W, lambda)
}
\arguments{
\item{W}{Spectral irradiance in W/m^2/nm.}

\item{lambda}{Wavelength(s) in nm.}
}
\value{
Numeric vector of photon counts per second per m^2 per nm.
}
\description{
Inverse of \code{\link{photon2W}}.
}
\examples{
W2photon(1, 555)
}
